+++ task-manager/app/add.tsx
@@ -1,8 +1,8 @@
 import { useRouter } from "expo-router";
 import { useState } from "react";
-import { KeyboardAvoidingView, Platform, ScrollView, StyleSheet, View } from "react-native";
-import { Button, Text, TextInput, RadioButton } from "react-native-paper";
+import { KeyboardAvoidingView, Platform, StyleSheet} from "react-native";
 import { Task } from "../types/task";
+import TaskForm from "../components/TaskForm";
 import { useTaskContext } from "../context/TaskContext";
 
 export default function AddTaskScreen() {
@@ -34,116 +34,17 @@ export default function AddTaskScreen() {
 			behavior={Platform.OS === "ios" ? "padding" : "height"}
 			style={{ flex: 1 }}
 		>
-			<ScrollView contentContainerStyle={styles.container}>
-				<View style={styles.headerContainer}>
-					<Text style={styles.headerText}>New Task</Text>
-				</View>
-
-
-				<View style={styles.inputGroup}>
-					<Text style={styles.label}>Title</Text>
-					<TextInput
-						mode="outlined"
-						placeholder="e.g., Grocery Shopping"
-						value={title}
-						onChangeText={setTitle}
-						style={styles.input}
-					/>
-				</View>
-
-
-				<View style={styles.inputGroup}>
-					<Text style={styles.label}>Description</Text>
-					<TextInput
-						mode="outlined"
-						placeholder="Details about the task..."
-						value={description}
-						onChangeText={setDescription}
-						multiline
-						numberOfLines={6}
-						style={styles.discrptiontextArea}
-					/>
-				</View>
-
-				<View style={styles.inputGroup}>
-					<Text style={styles.label}>Status</Text>
-					<RadioButton.Group onValueChange={value => setStatus(value as "pending" | "completed")} value={status}>
-						<View style={styles.radioRow}>
-							<RadioButton value="pending" />
-							<Text>Pending</Text>
-						</View>
-						<View style={styles.radioRow}>
-							<RadioButton value="completed" />
-							<Text>Completed</Text>
-						</View>
-					</RadioButton.Group>
-				</View>
-
-				{/* Save Button */}
-				<Button
-					mode="contained"
-					onPress={handleSave}
-					buttonColor="#8c7cde"
-					disabled={title.trim() === ""}
-					style={styles.submitButton}
-				>
-					Save
-				</Button>
-			</ScrollView>
+			<TaskForm
+        header="New Task"
+        title={title}
+        description={description}
+        status={status}
+        onChangeTitle={setTitle}
+        onChangeDescription={setDescription}
+        onChangeStatus={setStatus}
+        onSubmit={handleSave}
+      />
 		</KeyboardAvoidingView>
 	);
 }
 
-const styles = StyleSheet.create({
-	container: {
-		padding: 20,
-		flex: 1,
-		backgroundColor: "#fff",
-	},
-	headerContainer: {
-		backgroundColor: "#f6f2f9",
-		paddingVertical: 20,
-		paddingHorizontal: 20,
-		flexDirection: "row",
-		justifyContent: "center",
-		alignItems: "center",
-		position: "relative",
-		marginBottom: 40,
-		borderRadius: 16,
-	},
-
-	headerText: {
-		fontSize: 18,
-		fontWeight: "bold",
-		textAlign: "center",
-		flex: 1,
-	},
-
-	inputGroup: {
-		marginBottom: 25,
-	},
-	label: {
-		marginBottom: 10,
-		fontSize: 14,
-		fontWeight: "600",
-	},
-	input: {
-		backgroundColor: "#fff",
-	},
-	discrptiontextArea: {
-		backgroundColor: "#fff",
-		height: 160,
-		textAlignVertical: "top",
-	},
-	radioRow: {
-		flexDirection: "row",
-		alignItems: "center",
-		marginTop: 6,
-	},
-	submitButton: {
-		marginTop: 24,
-		paddingVertical: 6,
-		color: "#bb91de",
-	},
-})
-	;
